[chore] Home 화면 관련 코드 폴더링 작업
