+++ Logit/Logit/Home/ProjectListSection.swift
@@ -1,0 +1,126 @@
+//
+//  ProjectListSection.swift
+//  Logit
+//
+//  Created by 임재현 on 1/25/26.
+//
+
+import SwiftUI
+
+struct ProjectListSection: View {
+    let hasProjects: Bool
+    
+    var body: some View {
+        VStack(alignment: .leading, spacing: 8) {
+            // 헤더
+            HStack {
+                Text("프로젝트 목록")
+                    .typo(.body3_bold)
+                    .foregroundStyle(.black)
+                
+                Spacer()
+//
+//                if hasProjects {
+//                    Button {
+//                        // 전체보기 액션
+//                    } label: {
+//                        HStack(spacing: 4) {
+//                            Text("전체보기")
+//                                .typo(.body5_medium)
+//                                .foregroundStyle(.gray200)
+//
+//                            Image(systemName: "chevron.right")
+//                                .font(.system(size: 12))
+//                                .foregroundStyle(.gray200)
+//                        }
+//                    }
+//                }
+            }
+            .padding(.horizontal, 20)
+            
+            // 컨텐츠
+            if hasProjects {
+                ProjectListView()
+            } else {
+                ProjectEmptyView()
+            }
+        }
+    }
+}
+
+struct ProjectEmptyView: View {
+    var body: some View {
+        VStack {
+            Image("app_status_empty")
+                .resizable()
+                .scaledToFit()
+                .frame(width: 150,height: 117)
+        }
+        .frame(maxWidth: .infinity,maxHeight: .infinity)
+        .padding(.vertical, 60)
+        .background(.white)
+        .cornerRadius(16)
+        .padding(.horizontal, 20)
+    }
+}
+
+
+struct ProjectListView: View {
+    // 나중에 실제 데이터로 교체
+    let mockProjects = [
+        ("프로젝트 1", "2024.01.15"),
+        ("프로젝트 2", "2023.12.20"),
+        ("프로젝트 3", "2023.11.05"),
+        ("프로젝트 4", "2023.11.05"),
+        ("프로젝트 5", "2023.11.05")
+    ]
+    
+    var body: some View {
+        VStack(spacing: 0) {
+            ForEach(mockProjects.indices, id: \.self) { index in
+                ProjectCardCell(
+                    title: mockProjects[index].0,
+                    date: mockProjects[index].1
+                )
+                
+                // Divider (마지막 셀 제외)
+                if index < mockProjects.count - 1 {
+                    Divider()
+                        .background(Color.gray100)
+                        .padding(.horizontal, 20)
+                }
+            }
+        }
+    }
+}
+
+struct ProjectCardCell: View {
+    let title: String
+    let date: String
+    
+    var body: some View {
+        HStack(spacing: 12) {
+            // 세로 막대기 (태그)
+            RoundedRectangle(cornerRadius: 2)
+                .fill(Color.blue)
+                .frame(width: 3, height: 24)
+            
+            // 타이틀
+            Text(title)
+                .typo(.body6_medium)
+                .foregroundStyle(.black)
+                .lineLimit(1)
+            
+            Spacer()
+            
+            // 날짜
+            Text(date)
+                .typo(.body7_regular_140)
+                .foregroundStyle(.gray100)
+        }
+        .padding(.horizontal, 20)
+        .padding(.vertical, 17.5)
+        .background(Color.white)
+    }
+}
+
